backend/requirements.txt
fastapi
uvicorn
python-multipart
# google-generativeai # Using direct API for now or langchain
google-generativeai
pandas
regex
python-dotenv
